Rename dependencies.ts -> schemas.ts
And move all schemas into this file. Also define the sub-schemas as
actual zod schemas and infer their types.
